requirements.txt
fastapi
python-dotenv
uvicorn
firebase-admin
jinja2 # template engine
pydantic-settings